+++ 4eddit/src/Components/PostsCard/index.js
@@ -1,16 +1,24 @@
-import React from 'react'
+import React, { useEffect, useState} from 'react'
 import { CardContainer, InfoBars, ContentContainer, BottomDetails } from '../../Styled'
 import { ArrowUpward, ArrowDownward } from '@material-ui/icons'
 import { Typography } from '@material-ui/core'
 import { useHistory, useParams } from 'react-router-dom'
+import Skeleton from 'react-loading-skeleton'
 
 const PostCard = (props) => {
 
   const postInfo = props.post
-  
   const history = useHistory()
-
   const { postId } = useParams()
+  const [ isLoading, setIsLoading ] = useState(true)
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setIsLoading(false)
+    }, 2100)
+    return () => clearTimeout(timer)
+
+  },[])
 
   const goToDetails = (id) => {
     history.push(`/posts/${id}`)
@@ -19,20 +27,20 @@ const PostCard = (props) => {
 	return(
 		<CardContainer>
 			<InfoBars>
-        <Typography variant='h5'>{postInfo.username}</Typography>
+        <Typography variant='h5'>{ isLoading  ? <Skeleton height={20} width={200} /> : postInfo.username}</Typography>
 			</InfoBars>
 			<ContentContainer>
-        <Typography variant='h5'>{postInfo.title}</Typography>
-        <Typography variant='h6'>{postInfo.text}</Typography>
+        <Typography variant='h5'>{ isLoading  ? <Skeleton height={20} width={200} /> : postInfo.title}</Typography>
+        <Typography variant='h6'>{ isLoading  ? <Skeleton height={20} width={200} /> : postInfo.text}</Typography>
 			</ContentContainer>
       <InfoBars>
 				<BottomDetails>
 					<ArrowUpward />
-          <Typography>{postInfo.votesCount}</Typography>
+          <Typography>{ isLoading  ? <Skeleton height={16} width={20} /> : postInfo.votesCount}</Typography>
 					<ArrowDownward />
 				</BottomDetails>
 				<BottomDetails>
-          { postId ? <Typography>{postInfo.commentsCount} comentários</Typography>:
+          { postId ? <Typography>{ isLoading  ? <Skeleton height={16} width={20} /> : postInfo.commentsCount} comentários</Typography>:
           <Typography onClick={() => goToDetails(postInfo.id)}>{postInfo.commentsCount} comentários</Typography>
           }
 				</BottomDetails>
