+++ 4eddit/src/Pages/PostsPage/index.js
@@ -1,12 +1,13 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import { Container, GoBackContainer } from './Styled'
-import { useRequestDetails } from '../../Hooks/UseRequestDetails'
 import { useParams, useHistory } from 'react-router-dom'
 import PostCard from '../../Components/PostsCard'
 import CommentsCard from '../../Components/CommentsCard'
 import {  ArrowBack } from '@material-ui/icons'
 import {  makeStyles, Fab } from '@material-ui/core'
 import { FullContainer } from '../../Styled'
+import api from '../../Utils/Api/Api'
+import Skeleton from 'react-loading-skeleton'
 
 const useStyles = makeStyles((theme) => ({
   margin: {
@@ -17,13 +18,29 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
+const token = localStorage.getItem('token')
+
 const PostsPage = () => {
 
+  const [ details, setDetails ] = useState({})
   const classes = useStyles()
   const { postId } = useParams()
   const history = useHistory()
 
-  const { details } = useRequestDetails(`/posts/${postId}`)
+  const getDetails = async() => {
+    const axiosConfig = {
+      headers:  {
+          Authorization: token
+      }
+    }
+
+    const response = await api.get(`/posts/${postId}`, axiosConfig)
+    setDetails(response.data.post)
+  }
+
+  useEffect(() => {
+    getDetails()
+  }, [])
 
   const comments = details.comments
 
@@ -31,9 +48,8 @@ const PostsPage = () => {
     <FullContainer>
       <Container>
         <PostCard post={details} />
-        {comments && comments.map(comment => {
-          return <CommentsCard key={comment.id} comment={comment} />
-        })}
+        {comments === undefined ? Array(9).fill().map((item, idx) => (<Skeleton key={idx} width={600} height={200} />)) :
+          comments.map(comment => (<CommentsCard key={comment.id} comment={comment} />))}
         <GoBackContainer>
           <Fab size='large' variant='extended' className={classes.margin} onClick={() => history.push('/posts')}>
             <ArrowBack className={classes.extendedIcon} />
